+++ services/121-service/src/programs/dto/create-program-fsp-configuration.dto.ts
@@ -16,8 +16,8 @@ export class CreateProgramFspConfigurationDto {
   @ApiProperty({
     example: 'test_account',
     description:
-      'Should be string (for e.g. name=username) or array of strings (for e.g. name=columnsToExport)',
+      'Should be string (for e.g. name=username) or array of strings (for e.g. name=columnsToExport) or JSON object (for e.g name=displayName)',
   })
   @IsNotEmpty()
-  value: string;
+  value: string | string[] | Record<string, string>;
 }