+++ services/121-service/src/registration/registration-view.entity.ts
@@ -43,7 +43,7 @@ import { RegistrationEntity } from './registration.entity';
       .addSelect('registration.preferredLanguage', 'preferredLanguage')
       .addSelect('registration.inclusionScore', 'inclusionScore')
       .addSelect('fsp.fsp', 'financialServiceProvider')
-      .addSelect('fsp.fspDisplayNamePortal', 'fspDisplayNamePortal')
+      .addSelect('fsp.displayName', 'fspDisplayName')
       .addSelect('registration.paymentCount', 'paymentCount')
       .addSelect(
         'registration.maxPayments - registration.paymentCount',
@@ -103,7 +103,7 @@ export class RegistrationViewEntity {
   public financialServiceProvider: FspName;
 
   @ViewColumn()
-  public fspDisplayNamePortal: string;
+  public fspDisplayName: string;
 
   /** This is an "auto" incrementing field with a registration ID per program. */
   @ViewColumn()