+++ services/121-service/test/registrations/send-message-with-placeholder.test.ts
@@ -35,7 +35,7 @@ describe('Send custom message with placeholders', () => {
   it('should send message with placeholder values processed', async () => {
     // Arrange
     const message =
-      'This is a test message with {{namePartnerOrganization}} and {{paymentAmountMultiplier}} and {{fspDisplayNamePortal}} and {{fullName}}';
+      'This is a test message with {{namePartnerOrganization}} and {{paymentAmountMultiplier}} and {{fspDisplayName[registrationAh.preferredLanguage]}} and {{fullName}}';
 
     // Act
     await sendMessage(
@@ -71,7 +71,7 @@ describe('Send custom message with placeholders', () => {
     );
     processedMessage = processedMessage.replace(
       new RegExp('{{fspDisplayNamePortal}}', 'g'),
-      fspIntersolveJson.fspDisplayNamePortal,
+      fspIntersolveJson.displayName[registrationAh.preferredLanguage],
     );
     processedMessage = processedMessage.replace(
       new RegExp('{{fullName}}', 'g'),